autentikasi user dengan akun google

+++ database/migrations/2025_02_03_130948_user_volunteers.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('user_volunteers', function (Blueprint $table) {
             $table->increments('user_volunteer_id');
             $table->unsignedBigInteger('user_id'); // Tipe data sesuai dengan id pada tabel users
-            $table->unsignedInteger('volunteer_id'); // Tipe data disesuaikan dengan tabel volunteers
+            $table->unsignedBigInteger('volunteer_id'); // Tipe data disesuaikan dengan tabel volunteers
             $table->enum('status', ['pending', 'confirmed'])->default('pending');
             $table->timestamp('confirmed_at')->nullable();
